colexecjoin: extract HashJoinerSpec into colexecargs
This commit extracts `HashJoinerSpec` struct from `colexecjoin` into
`colexecargs` as well as exports a couple of its fields. This is needed
in order to reuse this struct by the hash group joiner that lives in
`colexec` package without having to make `colexec` depend on
`colexecjoin`.

Release note: None
