+++ pkg/sql/colexec/colexecjoin/BUILD.bazel
@@ -20,6 +20,7 @@ go_library(
         "//pkg/col/typeconv",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/colcontainer",
+        "//pkg/sql/colexec/colexecargs",
         "//pkg/sql/colexec/colexecbase",
         "//pkg/sql/colexec/colexechash",
         "//pkg/sql/colexec/colexecutils",
@@ -57,6 +58,7 @@ go_test(
         "//pkg/col/coldatatestutils",
         "//pkg/settings/cluster",
         "//pkg/sql/catalog/descpb",
+        "//pkg/sql/colexec/colexecargs",
         "//pkg/sql/colexec/colexectestutils",
         "//pkg/sql/colexecerror",
         "//pkg/sql/colexecop",
